+++ app/views/spam2/_revisions.html.erb
@@ -5,32 +5,39 @@ $(document).ready(function () {
 </script>
 
 <div class="card" id="table-card">
-  <div class="card-header">
+    <div class="bg-light navbar navbar-expand">
+      <ul class="nav navbar-expand navbar-nav-scroll">
+        <li class="nav-item">
+          <a class="btn nav-link text-secondary" data-toggle="tooltip" data-placement="top" title="Selected per page"> Selected <span id="select-count" class="badge badge-dark">0</span></a>
+        </li>
+      </ul>
   </div>
   <div class="card-body" style="overflow-x:hidden;">
+    <% unless @revisions.empty? %>
+      <div class="my-0 text-secondary small "><%= page_entries_info(@revisions) %>s</div>
+    <% end %>
     <table  id="revision_table" class="nowrap table table-hover" sstyle="width:100%; text-align:left">
       <thead style="text-align:left;">
         <tr>
           <th><input type="checkbox"  id="selectall" /></th>
           <th>Title</th>
           <th>Author</th>
-          <th>Time</th>
+          <th>Updated at</th>
           <th>Action</th>                                
         </tr>       
       </thead>
-
       <tbody> 
         <% @revisions.each do |revision| %>
           <tr id="n<%= revision.vid %>">
             <td><input  class="selectedId" value="<%= revision.vid %>" type="checkbox" /></td>
             <td>
-              <i class="fa fa-file"></i> <a class="text-primary" data-toggle="modal" data-target="#ninfo<%= revision.vid %>"><%= revision.title.truncate(20) %></a>
+              <i class="fa fa-file text-secondary"></i> <a class="text-dark font-weight-bold" data-toggle="modal" data-target="#ninfo<%= revision.vid %>" href="javascript:void(0);"> <%= revision.title.truncate(20) %></a><br />
             </td>
             <td>
-              <a href="/profile/<%= revision.author&.name %>"><%= revision.author&.name %></a> 
+              <a class="text-info" href="/profile/<%= revision.author&.name %>"><%= revision.author&.name %></a> 
             </td>
             <td>
-              <span style="color:grey;"><%= time_ago_in_words(revision.timestamp.seconds.ago) %> ago</span>
+              <span style="color:grey;"><%= time_ago_in_words(revision.updated_at) %> ago</span>
             </td>
             <td>
               <a class="badge badge-pill badge<% if revision.status == 0 %>-success<% else %>-secondary disabled<% end %> publish" data-remote="true" href="/moderate/revision/publish/<%= revision.vid %>"><i class="fa fa-check-circle fa-white"></i> Publish post</a>
@@ -65,7 +72,7 @@ $(document).ready(function () {
   <div class="float-right">
     <%= will_paginate @revisions, :renderer => WillPaginate::ActionView::BootstrapLinkRenderer unless @unpaginated || @revisions.empty?%>
   </div>
-  <div class="float-left">
-    Batch  <%= @revisions.current_page%> of <%= @revisions.total_pages %> total Batches
+  <div class="float-left text-secondary">
+    Batch  <%= @revisions.current_page%> of <%= @revisions.total_pages %> total Pages
   </div>
 </div>
