+++ config/routes.rb
@@ -297,14 +297,16 @@ Plots2::Application.routes.draw do
   get 'spam2/comments/filter/:type/:pagination' => 'spam2#_spam_comments'
   get 'spam2/users/filter/:type/:pagination' => 'spam2#_spam_users'
   get 'spam2/flags/filter/:type/:pagination' => 'spam2#_spam_flags'
+  get 'spam2/queue/filter/:tag' => 'spam2#_spam_queue'
   get 'spam2/filter/:type/:pagination' => 'spam2#_spam'
-  get 'spam2/batch_spam/:ids' => 'spam2#batch_spam'
-  get 'spam2/batch_publish/:ids' => 'spam2#batch_publish'
-  get 'spam2/batch_delete/:ids' => 'spam2#batch_delete'
-  get 'spam2/batch_ban/:ids' => 'spam2#batch_ban'
-  get 'spam2/batch_unban/:ids' => 'spam2#batch_unban'
-  get 'spam2/batch_ban_user/:ids' => 'spam2#batch_ban_user'
-  get 'spam2/batch_unban_user/:ids' => 'spam2#batch_unban_user'
+  get 'spam2/batch_spam/:ids' => 'batch#batch_spam'
+  get 'spam2/batch_publish/:ids' => 'batch#batch_publish'
+  get 'spam2/batch_delete/:ids' => 'batch#batch_delete'
+  get 'spam2/batch_ban/:ids' => 'batch#batch_ban'
+  get 'spam2/batch_unban/:ids' => 'batch#batch_unban'
+  get 'spam2/batch_ban_user/:ids' => 'batch#batch_ban_user'
+  get 'spam2/batch_unban_user/:ids' => 'batch#batch_unban_user'
+  get 'spam2/batch_comment/:type/:ids' => 'batch#batch_comment'
   get 'admin/users' => 'admin#users'
   get 'admin/queue' => 'admin#queue'
   get 'ban/:id' => 'admin#ban'
